pyproject.toml
[tool.poetry]
name = "send_email_python"
version = "0.1.0"
description = ""
authors = ["Ariulis <andrey59995@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
python-dotenv = "^0.20.0"
colorama = "^0.4.5"
tqdm = "^4.64.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.7.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
